Move functions from os/fs.c into path.c
Move unit tests from os/fs.moon to path.moon
