+++ src/syntax.c
@@ -34,6 +34,7 @@
 #include "garray.h"
 #include "option.h"
 #include "os_unix.h"
+#include "path.h"
 #include "regexp.h"
 #include "screen.h"
 #include "term.h"
