+++ test/unit/os/fs.moon
@@ -94,134 +94,6 @@ describe 'fs function', ->
       eq lfs.currentdir! .. '/unit-test-directory', (ffi.string buffer)
       eq OK, result
 
-  describe 'os_get_absolute_path', ->
-    ffi.cdef 'int os_get_absolute_path(char *fname, char *buf, int len, int force);'
-
-    os_get_absolute_path = (filename, buffer, length, force) ->
-      filename = to_cstr filename
-      fs.os_get_absolute_path filename, buffer, length, force
-
-    before_each ->
-      -- Create empty string buffer which will contain the resulting path.
-      export len = (string.len lfs.currentdir!) + 33
-      export buffer = cstr len, ''
-
-    it 'fails if given filename contains non-existing directory', ->
-      force_expansion = 1
-      result = os_get_absolute_path 'non_existing_dir/test.file', buffer, len, force_expansion
-      eq FAIL, result
-
-    it 'concatenates given filename if it does not contain a slash', ->
-      force_expansion = 1
-      result = os_get_absolute_path 'test.file', buffer, len, force_expansion
-      expected = lfs.currentdir! .. '/test.file'
-      eq expected, (ffi.string buffer)
-      eq OK, result
-
-    it 'concatenates given filename if it is a directory but does not contain a
-    slash', ->
-      force_expansion = 1
-      result = os_get_absolute_path '..', buffer, len, force_expansion
-      expected = lfs.currentdir! .. '/..'
-      eq expected, (ffi.string buffer)
-      eq OK, result
-
-    -- Is it possible for every developer to enter '..' directory while running
-    -- the unit tests? Which other directory would be better?
-    it 'enters given directory (instead of just concatenating the strings) if
-    possible and if path contains a slash', ->
-      force_expansion = 1
-      result = os_get_absolute_path '../test.file', buffer, len, force_expansion
-      old_dir = lfs.currentdir!
-      lfs.chdir '..'
-      expected = lfs.currentdir! .. '/test.file'
-      lfs.chdir old_dir
-      eq expected, (ffi.string buffer)
-      eq OK, result
-
-    it 'just copies the path if it is already absolute and force=0', ->
-      force_expansion = 0
-      absolute_path = '/absolute/path'
-      result = os_get_absolute_path absolute_path, buffer, len, force_expansion
-      eq absolute_path, (ffi.string buffer)
-      eq OK, result
-
-    it 'fails when the path is relative to HOME', ->
-      force_expansion = 1
-      absolute_path = '~/home.file'
-      result = os_get_absolute_path absolute_path, buffer, len, force_expansion
-      eq FAIL, result
-
-    it 'works with some "normal" relative path with directories', ->
-      force_expansion = 1
-      result = os_get_absolute_path 'unit-test-directory/test.file', buffer, len, force_expansion
-      eq OK, result
-      eq lfs.currentdir! .. '/unit-test-directory/test.file', (ffi.string buffer)
-
-    it 'does not modify the given filename', ->
-      force_expansion = 1
-      filename = to_cstr 'unit-test-directory/test.file'
-      -- Don't use the wrapper here but pass a cstring directly to the c
-      -- function.
-      result = fs.os_get_absolute_path filename, buffer, len, force_expansion
-      eq lfs.currentdir! .. '/unit-test-directory/test.file', (ffi.string buffer)
-      eq 'unit-test-directory/test.file', (ffi.string filename)
-      eq OK, result
-
-  describe 'append_path', ->
-    ffi.cdef 'int append_path(char *path, char *to_append, int max_len);'
-
-    it 'joins given paths with a slash', ->
-     path = cstr 100, 'path1'
-     to_append = to_cstr 'path2'
-     eq OK, (fs.append_path path, to_append, 100)
-     eq "path1/path2", (ffi.string path)
-
-    it 'joins given paths without adding an unnecessary slash', ->
-     path = cstr 100, 'path1/'
-     to_append = to_cstr 'path2'
-     eq OK, fs.append_path path, to_append, 100
-     eq "path1/path2", (ffi.string path)
-
-    it 'fails if there is not enough space left for to_append', ->
-      path = cstr 11, 'path1/'
-      to_append = to_cstr 'path2'
-      eq FAIL, (fs.append_path path, to_append, 11)
-
-    it 'does not append a slash if to_append is empty', ->
-      path = cstr 6, 'path1'
-      to_append = to_cstr ''
-      eq OK, (fs.append_path path, to_append, 6)
-      eq 'path1', (ffi.string path)
-
-    it 'does not append unnecessary dots', ->
-      path = cstr 6, 'path1'
-      to_append = to_cstr '.'
-      eq OK, (fs.append_path path, to_append, 6)
-      eq 'path1', (ffi.string path)
-
-    it 'copies to_append to path, if path is empty', ->
-      path = cstr 7, ''
-      to_append = to_cstr '/path2'
-      eq OK, (fs.append_path path, to_append, 7)
-      eq '/path2', (ffi.string path)
-
-  describe 'os_is_absolute_path', ->
-    ffi.cdef 'int os_is_absolute_path(char *fname);'
-
-    os_is_absolute_path = (filename) ->
-      filename = to_cstr filename
-      fs.os_is_absolute_path filename
-
-    it 'returns true if filename starts with a slash', ->
-      eq OK, os_is_absolute_path '/some/directory/'
-
-    it 'returns true if filename starts with a tilde', ->
-      eq OK, os_is_absolute_path '~/in/my/home~/directory'
-
-    it 'returns false if filename starts not with slash nor tilde', ->
-      eq FAIL, os_is_absolute_path 'not/in/my/home~/directory'
-
   describe 'os_isdir', ->
     os_isdir = (name) ->
       fs.os_isdir (to_cstr name)
